Moved tests to a different class.
